requirements.txt
annotated-types==0.7.0
anyio==4.12.0
argon2==0.1.10
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
bleach==6.2.0
blis==1.3.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cssselect2==0.8.0
cymem==2.0.13
Django==4.2.27
django-cors-headers==4.9.0
django-filer==3.4.1
django-filter==25.1
django-polymorphic==4.1.0
django-redis==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-spectacular==0.29.0
drf-yasg==1.21.11
easy-thumbnails==2.10.1
email-validator==2.3.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
eval_type_backport==0.3.1
exceptiongroup==1.3.1
google-auth==2.45.0
google-genai==1.47.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
invoke==2.2.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
mistralai==1.10.0
murmurhash==1.0.15
numpy==2.0.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pillow==11.3.0
preshed==3.0.12
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
reportlab==4.4.7
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
svglib==1.5.1
tenacity==9.1.2
thinc==8.3.10
tinycss2==1.4.0
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
wasabi==1.1.3
weasel==0.4.3
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.1
zipp==3.23.0
